Organize utility package
