+++ src/main/java/com/meti/app/ClientDisplay.java
@@ -4,8 +4,8 @@ import com.meti.lib.fx.Controller;
 import com.meti.lib.fx.PostInitializable;
 import com.meti.lib.net.client.Client;
 import com.meti.lib.net.command.GetCommand;
-import com.meti.lib.util.CollectionUtil;
-import com.meti.lib.util.StringConverter;
+import com.meti.lib.util.collect.CollectionUtil;
+import com.meti.lib.util.convert.StringConverter;
 import javafx.fxml.FXML;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeView;
