+++ src/main/java/com/meti/lib/util/collect/StreamUtil.java
@@ -1,4 +1,4 @@
-package com.meti.lib.util;
+package com.meti.lib.util.collect;
 
 import java.util.List;
 import java.util.Set;
