+++ src/main/java/com/meti/lib/util/convert/ClassConverter.java
@@ -1,4 +1,4 @@
-package com.meti.lib.util;
+package com.meti.lib.util.convert;
 
 /**
  * @author SirMathhman
