+++ src/main/java/com/meti/lib/util/convert/StringConverter.java
@@ -1,4 +1,4 @@
-package com.meti.lib.util;
+package com.meti.lib.util.convert;
 
 /**
  * @author SirMathhman
